Fix building on arm64

+++ mailbox.c
@@ -61,7 +61,7 @@ void *mapmem(unsigned base, unsigned size)
    printf("base=0x%x, mem=%p\n", base, mem);
 #endif
    if (mem == MAP_FAILED) {
-      printf("mmap error %d\n", (int)mem);
+      printf("mmap error %ld\n", (long)mem);
       exit (-1);
    }
    close(mem_fd);
